requirements.txt
discord
flask
sqlalchemy
eralchemy
psycopg2-binary
alembic
pyyaml
requests
flask-wtf
Flask-Misaka
click